% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{s1}
\alias{s1}
\title{Gas Production Data (s1)}
\format{
A data frame with 31 rows and 11 columns.
}
\source{
Original data was sourced from [https://www.aer.ca/providing-information/data-and-reports/statistical-reports/st98/statistics-and-data].
}
\description{
This dataset contains annual gas production data in million cubic meters per day.
The data was extracted from the 'ST98-NaturalGas-SupplyDemand-data.xlsx' Excel file.
}
