% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{s10}
\alias{s10}
\title{Gas Purchases Data (s10)}
\format{
A data frame with 33 rows and 5 columns.
}
\source{
Original data was sourced from [https://www.aer.ca/providing-information/data-and-reports/statistical-reports/st98/statistics-and-data].
}
\description{
This dataset contains number of horizontal and vertical wells as well as production in million cubic meters per day.
The data was extracted from the 'ST98-NaturalGas-SupplyDemand-data.xlsx' Excel file.
}
