% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{s2p}
\alias{s2p}
\title{Wells Placed On Production Data, Horizontal and Vertical (s2p)}
\format{
A data frame with 58 rows and 3 columns.
}
\source{
Original data was sourced from [https://www.aer.ca/providing-information/data-and-reports/statistical-reports/st98/statistics-and-data].
}
\description{
This dataset contains annual wells placed on production data, as well as marketable production in million cubic meters per day.
The data was extracted from the 'ST98-NaturalGas-SupplyDemand-data.xlsx' Excel file.
}
