% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{s3}
\alias{s3}
\title{Bitumen Well Data (s3)}
\format{
A data frame with 33 rows and 3 columns.
}
\source{
Original data was sourced from [https://www.aer.ca/providing-information/data-and-reports/statistical-reports/st98/statistics-and-data].
}
\description{
This dataset contains annual bitumen well data in million cubic meters per day.
The data was extracted from the 'ST98-NaturalGas-SupplyDemand-data.xlsx' Excel file.
}
