% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{s5}
\alias{s5}
\title{Shale Well Data (s5)}
\format{
A data frame with 33 rows and 4 columns.
}
\source{
Original data was sourced from [https://www.aer.ca/providing-information/data-and-reports/statistical-reports/st98/statistics-and-data].
}
\description{
This dataset contains annual shale well data in number of wells, as well as Alberta Plant gate price in Cdn$/GJ.
The data was extracted from the 'ST98-NaturalGas-SupplyDemand-data.xlsx' Excel file.
}
