% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{s7}
\alias{s7}
\title{Gas Demand Data (s7)}
\format{
A data frame with 33 rows and 10 columns.
}
\source{
Original data was sourced from [https://www.aer.ca/providing-information/data-and-reports/statistical-reports/st98/statistics-and-data].
}
\description{
This dataset contains annual gas demand  by source in million cubic meters per day.
The data was extracted from the 'ST98-NaturalGas-SupplyDemand-data.xlsx' Excel file.
}
