% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{s9}
\alias{s9}
\title{Gas Purchases Data (s9)}
\format{
A data frame with 23 rows and 2 columns.
}
\source{
Original data was sourced from [https://www.aer.ca/providing-information/data-and-reports/statistical-reports/st98/statistics-and-data].
}
\description{
This dataset contains annual land available for permitting, in billion cubic meters.
The data was extracted from the 'ST98-NaturalGas-SupplyDemand-data.xlsx' Excel file.
}
